% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftplot.R
\name{ftplot}
\alias{ftplot}
\title{Easy plot highcharter personal theme}
\usage{
ftplot(hc, title, subtitle, source, range = TRUE, navigator = TRUE)
}
\arguments{
\item{hc}{highchart object}

\item{title}{title to the plot}

\item{subtitle}{subtitle to the plot}

\item{source}{source of data}

\item{range}{logical, display range options?}

\item{navigator}{logical, display navigator zoom?}
}
\value{
plot
}
\description{
Easy plot highcharter personal theme
}
