% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipea.R
\name{get_ipea}
\alias{get_ipea}
\title{Download series from IPEADATA}
\usage{
get_ipea(code, silent = FALSE)
}
\arguments{
\item{code}{IPEADATA code}

\item{silent}{Logical}
}
\value{
Tibble
}
\description{
Download series from IPEADATA
}
