% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFedData.R
\name{getFedData}
\alias{getFedData}
\title{Title
Fetches and returns clean data from the Federal Reserces API using the FedR package}
\usage{
getFedData(tag = "DFF")
}
\arguments{
\item{tag}{}
}
\value{
dt
}
\description{
Title
Fetches and returns clean data from the Federal Reserces API using the FedR package
}
\examples{
fed_functs <- getFedData("DFF")
}
