% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_sparkline.R
\name{gt_sparkline}
\alias{gt_sparkline}
\title{Create sparkline in gt object}
\usage{
gt_sparkline(table_data, plot_col, data_col)
}
\arguments{
\item{table_data}{tibble}

\item{plot_col}{plot_col}

\item{data_col}{data_col}
}
\value{
spartline col
}
\description{
Create sparkline in gt object
}
