% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry an expression}
\usage{
retry(foo, tries = 5)
}
\arguments{
\item{foo}{The expression}

\item{tries}{Number of tries}
}
\value{
Output of expression or error
}
\description{
Retry an expression
}
